Python_function_Exercises_27

+++ Python-Exercises/Python_P_Conditions_Loops_Ex_13092025/Lab062_Python_Triangle_Classifier.py
@@ -8,9 +8,9 @@
 # user i/p - s1,s2,s3 - int
 # o/p - iso, eq, scelene
 
-s1 = int(input("Enter a first side length: "))
-s2 = int(input("Enter a second side length: "))
-s3 = int(input("Enter a third side length: "))
+s1 = float(input("Enter a first side length: "))
+s2 = float(input("Enter a second side length: "))
+s3 = float(input("Enter a third side length: "))
 
 if s1==s2 and s1==s3:
     print("All sides are equal, Hence, it's a equilateral triangle")
